added confirm dialog

+++ frontend/src/app/books/container/book-details/book-details.component.ts
@@ -27,6 +27,10 @@ export class BookDetailsComponent implements OnInit {
   }
 
   deleteBook(book: Book) {
+    if (!confirm('really delete')) {
+      return;
+    }
+
     this.bookService.delete(book.id).subscribe(() => {
       this.router.navigate(['/books/overview/']);
     });
